LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/property-postmessage-2.https-expected.txt

CONSOLE MESSAGE: Unable to post message to http://web-platform.test:8800. Recipient has origin https://web-platform.test:9443.

CONSOLE MESSAGE: Unable to post message to http://web-platform.test:8800. Recipient has origin https://www1.web-platform.test:9443.


FAIL same-origin > w => w.postMessage("") promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'report.body.property')"
FAIL cross-origin > w => w.postMessage("") promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'report.body.property')"
